add append, prepend, size methods
Probably should have committed these one at a time, oh well! append adds
a node to the end of the list, prepend adds it to the start of the list,
and size returns the total size of the list. I added some testing stuff
to the index.js file so that when I run it I can see if my methods are
working correctly. (TDD lite?)