Added getHead method to return the head of list
This may be cheating because the assignment says to call this method
head, but personally I just feel like I would never live in a world
where a method would have the same name as a property of an object or
class. It would HAVE to be called getHead because it is both humorous
and also becasue namespace pollution and all that jazz.